reorganize run scripts
